requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
gradio==6.0.0
pymongo==4.15.4
python-dotenv==1.2.1
httpx==0.28.1
google-generativeai==0.8.5



